MINOR: Reorganise scaffolding tests
